Update PHPUnit requirement to ^8.0 || ^9.0
In order to ensure the code works on PHP 8 as well the inspections should also run on PHP 8. As the library runs perfectly fine on PHP 7.2 and PHP 7.3 I don't see a necessity to drop support for those versions. In order to still support PHP 8 PHPUnit needs to be updated to version 9, but that version drops support for PHP 7.2 and PHP 7.3. The solution is to require either PHPUnit 8 or PHPUnit 9.

This left an issue with the usage of Prophecy as Prophecy is deprecated in PHPUnit 9. As the test suite is so small the solution for this project has been to add `phpspec/prophecy` as a direct development dependency and using Prophecy direct instead of via `\PHPUnit\Framework\TestCase::prophesize()`.